--- python_folder/README.md
@@ -1,1 +1,0 @@
-# legal-text-summarizer
