--- python_folder/dataset.py
@@ -1,3 +1,0 @@
-from datasets import load_dataset
-dataset = load_dataset("ninadn/indian-legal")
-print(dataset)
