+++ python_folder/test_files/bert.py
@@ -1,0 +1,6 @@
+from summarizer import Summarizer,TransformerSummarizer
+legal_text = '''
+1 IN THE COURT OF SESSIONS, ERNAKULAM DIVISION Present: Smt. Reshma Sasidharan, II Addl. Assistant Sessions Judge, Ernakulam. Thursday, the 24th day of January, 2025/ 4th Magha, 1946 SESSIONS CASE No.210/2023 (C.P. No.26/2022 of Additional Chief Judicial Magistrate (EO), Ernakulam Crime No.55/2022 of Excise Range Office, Ernakulam) Complainant : State of Kerala, Rep. by the Excise Circle Inspector, Ernakulam (By Adv. Smt. Suja T Addl. Public Prosecutor,) Accused:- Raveendran, S/o Narayanan, aged 63/2022, Madassery House, Pandit Karuppan Road, Thevara College Desom, Ernakulam Village, Kanayannur Taluk, Ernakulam District. (By Adv. Manu Roy) Offence : Under Section 55 (i) and 58 of Abkari Act 1 of 1077 Plea of the accused : Not guilty Finding of the court : Not guilty Sentence or order : The accused is found not guilty of the offences punishable under Section 55 (i) and Section 58 of the Abkari Act and he is acquitted under Section 235 (1) of Cr.PC. The accused is set at liberty and his bail bond stands cancelled. 2 DESCRIPTION OF THE ACCUSED Sl. No . Name Father’s Name Occupation Age Residence Raveendran Narayanan -- 63/22 Madassery House, Pandit Karuppan Road, Thevara College Desom, Ernakulam Village, Kanayannur Taluk, Ernakulam District. Date of Occurrence : 29.07.2022 Final Report : 30.11.2022 Apprehension : 29.07.2022 Release on bail : 19.08.2022 Committal for trial : 04.01.2023 Commencement of trial : 23.09.2023 Close of trial : 17.01.2025 Sentence or order : 24.01.2025 The above case was committed to the Hon’ble Court of Session Ernakulam, Division by the Judicial First Class Magistrate-II, Ernakulam. As per order dated 04.01.2023 in Committal Proceedings No. 26/2022 and thereafter made over to this court for trial and disposal. 3 The above case having been finally heard on 20.01.2025 in the presence of the learned Additional Public Prosecutor and the learned counsel for the accused, and having stood over for consideration to this day, the court on 24.01.2025 delivered the following :- JUDGMENT This is a case which arose out of a complaint filed by the Excise Inspector, Ernakulam in crime No.55/2022 alleging commission of the offence punishable under Section 55(i ) , 58 of the Abkari Act. 2. The prosecution case is that on 29.07.2022 at 12.45 p.m the accused was found to be in possession of 18.5 litre of Indian made foreign liquor for the purpose of sale in a lottery shop situated near Chakkola junction, Thevara- ferry road, Ernakulam. Thus the accused is alleged to have committed the aforesaid offence. 3. On filing the final report before the Additional Chief Judicial Magistrate (EO) , Ernakulam, the case was taken on file as CP No.26/2022. After complying with Section 207 of Cr.PC the learned Magistrate committed the case to the Hon’ble Court of Sessions, Ernakulam. Later the case is made over to this Court for trial and disposal. The accused is on bail from the crime stage. 4 4. After hearing both sides charge is framed for the above offences read over and explained to the accused to which he pleaded not guilty and claimed to be tried. Thereafter the prosecution examined PWs 1 to 6 and marked Exts. P1 to P9. 5. After closing the prosecution evidence the accused is examined under Section 313 1(b) of Cr.PC and he denied all the incriminating circumstances appearing in evidence against him and maintained innocence. Since there was no ground for an acquittal under Section 232 of Cr.PC the accused was directed to enter upon his defence. But no defence evidence was adduced. 6. Heard both sides. 7. The points that arise for consideration are:- 1. Has the prosecution proved that the accused on 29.07.2022 at 12.45 p.m in his lottery shop situated at Chakkola junction was engaged in the sale of illicit liquor as alleged? 2. Has the prosecution proved that the accused on the aforementioned date and at the above mentioned place 5 was found in possession of illicit liquor knowing the same to have been unlawfully imported, transported or manufactured, or knowing the duty, tax or rent payable under the Abkari Act not to have been paid? 3. Whether the accused is guilty or not? 4. Sentence or order? Point Nos.1 and 2 8. The prosecution is relying on the oral testimony of PW1, the detecting officer and Exts. P3, P6 and P9 and the chemical analysis report to prove the alleged occurrence. PW1 was the Excise Range Officer, Ernakulam. He along with Excise party reached the place of occurrence in connection with the information received by them. At that time the accused was found engaged in the sale of Indian made foreign liquor. The person to whom the accused sold the liquor was not caught and the glass in which the liquor was served was also not recovered from the place of occurrence. No bottle of liquor was seized from the immediate possession of the accused. The accused is a lottery vendor and he has placed the lotteries for sale in front of a shop near the Chakkola junction. The shop room is actually rented out to one Chellappan, who is also arrayed as a 6 witness in this case, by one Anna John Chakkola. It is alleged that without any written agreement the said Chellappan sublet the portion in front of the shop to the accused for selling lottery tickets. The 18.5 litre of Indian made foreign liquor were found on the south-western corner of the shop room under the washbasin inside a sack. The bottle from which the accused was found serving liquor to the buyer was also not found out. After being convinced of the identity of the accused he was arrested by PW1. PW1 proved the Ext.P1 arrest memo and Ext.P2 intimation memo. It is to be noted that no thondy money was recovered from the possession of the accused. PW1 took the 18.5 litre of liquor bottles into custody , sealed and labelled the same. He drew samples from the bottles and sealed and labelled the same also. PW1 proved the Ext.P3 as the seizure mahazar prepared by him. He produced the accused , the thondy materials and the case records at the Excise Range Office. PW1 identified the accused standing in the dock. He denied the suggestion that nothing is recovered from the immediate possession of the accused and that the accused was falsely implicated in this case. 9. PW2 was the Civil Excise Officer and he accompanied PW1 during the patrolling duty and also assisted PW1 in the seizure of liquor and he also assisted PW1 in preparing the case records. He identified the 7 signature in Ext.P3 and also identified the inventory prepared in this case. He identified the accused standing in the dock. He denied the suggestion that the accused is in no way connected with the shop room from where the liquor was seized and also denied that a false case is filed against this accused. 10. PW3 is a relative of the landlord Anna John Chakkola, who let the shop room to CW5 , Chellappan. PW3 deposed that upon getting information about the seizure when he reached there he saw the excise officials transporting the liquor bottles in a box and loading it in the Excise vehicle. He is a witness to Ext.P3 seizure mahazar and also to the Ext P1 arrest memo. He identified the accused standing in the dock. During his cross examination PW3 stated that he saw the Excise officials putting the liquor bottles in a box and keeping it inside the official vehicle of the Excise Officer. He further deposed that when he reached there most of the bottles were already shifted to the vehicle and he saw the Excise officials taking the remaining bottles kept under the wash basin in the shop room and transferring it to the jeep. He also deposed that CW5 Chellappan is the tenant . PW3 also did not depose that the liquor was seized from the immediate possession of the accused and that he had any previous knowledge about the alleged illegal sale of liquor conducted by the 8 accused. 11. PW5 was the Excise Range Inspector , Ernakulam and he registered the crime and occurrence report Ext.P5 in this case. PW5 proved the Ext.P6 property list and the Ext.P7 remand report. He identified the sample seal Ext.P8 produced before the Court . He also proved the Ext P9 inventory report. PW5 identified the accused standing in the dock. 12. PW4 proved the Ext.P4 site plan. PW6 was the Excise Circle Inspector , Ernakulam and he conducted the investigation in this case. PW6 admitted that the person who bought liquor from accused was not caught. He also admitted that the CCTV footage of the area is not produced to prove that the accused was engaged in the illegal sale of liquor. PW6 did not go through the rent documents of the shop rooms and also did not produce the rent agreement between Anna John and CW5 Chellappan. He admitted that there is only an oral arrangement between the accused and the said Chellappan. He only understood that the accused was a lottery vendor and he has not perused any document to show the relationship of the accused with the shop room from where the liquor was seized. He also admitted that the neighbouring shop owners were not 9 enquired with the matter and no body was made a witness to the alleged illegal sale of the accused. He denied the suggestion that without conducting proper investigation a false case is lodged against the accused. 13. The only evidence available in this case to prove the occurrence is the oral testimony of PW1 who detected the crime, PW2 who accompanied PW1 and PW5 and PW6 who registered the crime and occurrence report and conducted investigation in this case. Even though PW3 testified about the occurrence, arrest of the accused and seizure of the liquor bottles there is no other material to substantiate the same. The detecting officer PW1 was equivalent to a dumb witness who does not know anything about the alleged seizure and arrest of the accused. The investigating officers also did not produce any evidence to connect the accused with the shop room from where the liquor was seized. There is no evidence to show that that portion of the shop room from where the liquor was seized was in the possession or control of the accused or that he was using that portion in connection with his lottery business. Even though PW1 alleges that he saw a person drinking liquor which was served by the accused he could not book the buyer and also could not recover the bottles from which the accused was serving liquor and the glass from which the buyer was drinking liquor from the premises. PW1 simply stated that the 10 buyer abandoned the glass and ran away from the place seeing the Excise jeep. But the glass used by the buyer which was allegedly abandoned also was not recovered and produced for proving the allegation. From the deposition of PW1 it is not clear whether the place from where the contraband were seized was a place where the accused had access. The lottery base of the accused was placed in the front portion of the shop room alone could be inferred from the prosecution case. The seizure of the liquor bottles from the shop room which is let out to CW5, Chellapppan by one Anna John, whether was sublet to the accused is not clear from the evidence adduced. The prosecution also failed to produce witness CW5 Chellappan to prove that there was an oral agreement between the said Chellappan and the accused to use the front portion of the shop room as well as the shop room in connection with the lottery business of the accused. PW 1 simply stated that there is only an oral agreement between the accused and said Chellappan. The Seizure was not done from the immediate possession of the accused . 14. The crime and occurrence report was registered against the accused under Section 55 (i ) and Section 58 of the Abkari Act. Section 55 (I) of the Abkari Act states that who ever in contravention of this Act sells or stores for sale liquor or any intoxicating drug shall be punished. In the 11 present case the sale of liquor is not proved by the prosecution. Secondly the storing of the liquor by the accused is also not proved. There is no evidence to show that the shop room was in the immediate possession of the accused and that he stored the liquor inside the shop room for the purpose of sale. The possession contemplated under this section is conscious possession where the accused should be aware of the particular fact. It is the duty of the prosecution to prove that the accused was in immediate conscious possession of the contraband when the crime was detected. In other words, the prosecution must prove that the accused was knowingly in control of the sale. Only when possession is established, the person who claims that it was not a conscious possession, has to establish it. Here the prosecution did not prove the immediate and conscious possession of the contraband by the accused. 15. The other provision under which the accused is implicated is under Section 58 of the Abkari Act. Section 58 of the Abkari Act states that whoever , without lawful authority , has in his possession any quantity of liquor or of any intoxicating drug , knowing the same to have been unlawfully imported, transported or manufactured or knowing that the duty or tax payable under this Act is not paid. The possession contemplated under Section 58 of the Abkari Act is the stage after import, 12 export , transport etc. that had already taken place some time back at the hands of somebody else. Here the possession must be with the knowledge that the contraband is illegally imported or exported or transported. When the possession itself is not proved by the prosecution then the question whether the seized contraband were illegally imported, exported or transported does not arise. It is true that the prosecution could establish that the samples sent for analysis contained 38.89 %, 41.90 % , 41.56 % , 38.77 % and 42.18 % by volume of ethyl alcohol. But there is no evidence to show that the samples drawn and produced before the Court reached the chemical examination lab in a tamper proof manner. The prosecution should have examined the clerk concerned who received the sample in the Court and forwarded the same to the lab to prove the above fact. The same is also fatal to the prosecution case. 16. In order to bring home an indictment under Section 55 (i) of the Kerala Abkari Act, it must be established that the accused was engaged in the sale of liquor or that he stored the same for sale. It is the admitted case of the prosecution that the person who was present at the spot to take liquor from the accused ran away from there. But no attempt is made by PW1 or PW5 and PW6 to identify the said person or to cite him as a witness to prove the alleged sale of liquor. The prosecution has no 13 case that the accused handed over the liquor to the vendee for money. 17. From the above discussion, the Court is of the view that the prosecution failed to prove the guilt of the accused beyond reasonable doubt. As there is an element of doubt in the prosecution case , it goes in favour of the accused. Point No.3 Since the accused is found not guilty point No.4 need not be answered. In view of the findings on point No.1 and 2 the accused is found not guilty of the offences punishable under Section 55 (i) and Section 58 of the Abkari Act and he is acquitted under Section 235 (1) of Cr.PC. The accused is set at liberty and his bail bond stands cancelled. Dictated to the Confidential Assistant, typed by him, corrected and pronounced by me in the open court on this the 24th day of January, 2025 Sd/- Reshma Sasidharan, II Additional Assistant Sessions Judge APPENDIX Prosecution Witnesses:- PW1 Ratheesh T.K 11.12.2024 PW2 Tomy N.D 11.12.2024 PW3 Paul Thomma Chakola 12.12.2024 PW4 V.C.Rajendran 12.12.2024 PW5 Haneefa M.S 06.01.2025 14 PW6 Prince Babu 06.01.2025 Prosecution Exhibits:- P1 Arrest Memo Proved through PW1 29.07.2022 P2 Arrest Intimation Proved through PW1 29.07.2022 P3 Seizure Mahazar Proved through PW1 29.07.2022 P4 Site plan Proved through PW4 26.10.2022 P5 Crime and occurrence report Proved through PW5 29.07.2022 P6 Property list Proved through PW5 29.07.2022 P7 Remand report Proved through PW5 30.07.2022 P8 Specimen Seal Proved through PW5 -- P9 Inventory report Proved through PW5 -- Defence Witnesses:- Nil Material Objects:- Nil Id/- II Additional Assistant Sessions Judge //True copy// II Additional Assistant Sessions Judge 15 JUDGMENT in SC No.210/2023 Dt. 24.01.2025 '''
+bert_model = Summarizer()
+bert_summary = bert_model(legal_text, num_sentences=10)
+print(bert_summary)
