+++ python_folder/test_files/dataset.py
@@ -1,0 +1,3 @@
+from datasets import load_dataset
+dataset = load_dataset("ninadn/indian-legal")
+print(dataset)
